requirements.txt
gradio==4.44.0
hydra-core==1.3.2
omegaconf==2.3.0
pandas==2.2.2
requests==2.32.3
ollama==0.3.3
openpyxl==3.1.5
